pyproject.toml
[project]
name = "flow-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Sanket Patel",email = "44511437+s4nk37@users.noreply.github.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi (>=0.121.2,<0.122.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "pydantic[email] (>=2.12.4,<3.0.0)",
    "alembic (>=1.17.2,<2.0.0)",
    "passlib[argon2] (>=1.7.4,<2.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pydantic-settings (>=2.12.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "black (>=25.11.0,<26.0.0)",
    "isort (>=7.0.0,<8.0.0)",
    "pytest (>=9.0.1,<10.0.0)",
    "httpx (>=0.28.1,<0.29.0)"
]
